use AdapterDelegatee to provid footer
