+++ linen-modern/src/main/scala/x7c1/linen/modern/display/settings/ChannelRowAdapter.scala
@@ -3,38 +3,38 @@ package x7c1.linen.modern.display.settings
 import android.support.v7.widget.RecyclerView.Adapter
 import android.view.ViewGroup
 import x7c1.linen.glue.res.layout.{SettingMyChannelRow, SettingMyChannelRowFooter, SettingMyChannelRowItem}
-import x7c1.linen.modern.accessor.AccountIdentifiable
-import x7c1.linen.modern.accessor.setting.{MyChannelAccessor, SettingMyChannel, SettingMyChannelFooter}
-import x7c1.linen.modern.init.settings.my.MyChannelRowProviders
+import x7c1.linen.modern.accessor.setting.{SettingMyChannel, SettingMyChannelFooter}
+import x7c1.linen.modern.accessor.{AccountIdentifiable, setting}
+import x7c1.wheat.lore.resource.AdapterDelegatee
 import x7c1.wheat.modern.decorator.Imports._
 
 class ChannelRowAdapter(
-  accessor: MyChannelAccessor,
-  providers: MyChannelRowProviders,
+  accountId: Long,
+  delegatee: AdapterDelegatee[setting.SettingMyChannelRow, SettingMyChannelRow],
   onSourcesSelected: ChannelSourcesSelected => Unit,
   onSubscriptionChanged: MyChannelSubscriptionChanged => Unit
 ) extends Adapter[SettingMyChannelRow]{
 
-  override def getItemCount: Int = accessor.length
+  override def getItemCount: Int = delegatee.count
 
   override def onCreateViewHolder(parent: ViewGroup, viewType: Int) = {
-    providers.createViewHolder(parent, viewType)
+    delegatee.createViewHolder(parent, viewType)
   }
   override def onBindViewHolder(holder: SettingMyChannelRow, position: Int): Unit = {
-    accessor.bindViewHolder(holder, position){
+    delegatee.bindViewHolder(holder, position){
       case (holder: SettingMyChannelRowItem, channel: SettingMyChannel) =>
         holder.name.text = channel.name
         holder.description.text = channel.description
         holder.sources onClick { _ =>
           onSourcesSelected apply ChannelSourcesSelected(
-            accountId = accessor.accountId,
+            accountId = accountId,
             channelId = channel.channelId,
             channelName = channel.name
           )
         }
         holder.switchSubscribe onChangedManually { e =>
           onSubscriptionChanged apply MyChannelSubscriptionChanged(
-            accountId = accessor.accountId,
+            accountId = accountId,
             channelId = channel.channelId,
             isSubscribed = e.isChecked
           )
@@ -45,9 +45,7 @@ class ChannelRowAdapter(
         // nop
     }
   }
-  private lazy val viewTypeAt = providers createViewTyper accessor
-
-  override def getItemViewType(position: Int) = viewTypeAt(position)
+  override def getItemViewType(position: Int) = delegatee.viewTypeAt(position)
 }
 
 case class ChannelSourcesSelected(
